backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptg==0.5.2
cryptography==46.0.5
decorator==4.4.2
ecdsa==0.19.1
fastapi==0.133.1
googletrans==4.0.2
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
instagrapi==2.3.0
instaloader==4.15.1
moviepy==1.0.3
numpy==2.4.4
passlib==1.7.4
pillow==12.2.0
proglog==0.1.12
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyWavelets==1.9.0
PyYAML==6.0.3
rarfile==4.2
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
starlette==0.52.1
Telethon==1.42.0
telethon-secret-chat==0.2.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
pyotp==2.9.0
